add some improvements after review
