+++ FakeNFT/Foundation/Helpers/UIBlockingProgressHUD.swift
@@ -8,7 +8,6 @@
 import UIKit
 import ProgressHUD
 
-// MARK: - UIBlockingProgressHUD
 final class UIBlockingProgressHUD {
 
     private static var window: UIWindow? {
