+++ FakeNFT/Scenes /WebView/WebViewModel.swift
@@ -6,6 +6,7 @@
 //
 
 // MARK: - WebViewModelProtocol
+
 protocol WebViewModelProtocol: AnyObject {
     var onProgressChange: ((Float) -> Void)? { get set }
     var onProgressHide: ((Bool) -> Void)? { get set }
@@ -13,13 +14,16 @@ protocol WebViewModelProtocol: AnyObject {
 }
 
 // MARK: - WebViewModel
+
 final class WebViewModel: WebViewModelProtocol {
 
     // MARK: - Public properties
+
     var onProgressChange: ((Float) -> Void)?
     var onProgressHide: ((Bool) -> Void)?
     
     // MARK: - Public Methods
+
     func didUpdateProgressValue(_ newValue: Double) {
         let newProgressValue = Float(newValue)
         onProgressChange?(newProgressValue)
@@ -27,6 +31,7 @@ final class WebViewModel: WebViewModelProtocol {
     }
 
     // MARK: - Private Methods
+    
     private func shouldHideProgress(for value: Float) -> Bool {
         abs(value - 1.0) <= 0.0001
     }
